+++ src/main/java/com/rafael/helpdesk/domain/model/Tecnico.java
@@ -27,7 +27,7 @@ public class Tecnico extends Pessoa {
 		this.perfis.add(Perfil.TECNICO);
 	}
 
-	public Tecnico(Integer id, String nome, String cpf, String email, String senha) {
+	public Tecnico(Long id, String nome, String cpf, String email, String senha) {
 		super(id, nome, cpf, email, senha);
 		this.perfis.add(Perfil.TECNICO);
 	}
